fix: add dnd cypress stub
