+++ packages/visual-editor/src/components/RootRenderer/RootRenderer.tsx
@@ -1,22 +1,17 @@
 import React from 'react';
 import { useEffect } from 'react';
-import { DragDropContext } from '@hello-pangea/dnd';
 import { Dropzone } from '../Dropzone/Dropzone';
 import DraggableContainer from '../Draggable/DraggableComponentList';
-import { sendMessage } from '@contentful/experience-builder-core';
 import type { CompositionTree } from '@contentful/experience-builder-core/types';
-import { OUTGOING_EVENTS } from '@contentful/experience-builder-core/constants';
-import dragState from '@/utils/dragState';
-import { usePlaceholderStyle } from '@/hooks/usePlaceholderStyle';
+
 import { ROOT_ID } from '@/types/constants';
 import { useTreeStore } from '@/store/tree';
 import { useDraggedItemStore } from '@/store/draggedItem';
-import { useEditorStore } from '@/store/editor';
 import { useZoneStore } from '@/store/zone';
 import styles from './render.module.css';
 import { useBreakpoints } from '@/hooks/useBreakpoints';
 import { useEditorSubscriber } from '@/hooks/useEditorSubscriber';
-import useCanvasInteractions from '@/hooks/useCanvasInteractions';
+import { DNDProvider } from './DNDProvider';
 
 interface Props {
   onChange?: (data: CompositionTree) => void;
@@ -25,9 +20,7 @@ interface Props {
 export const RootRenderer: React.FC<Props> = ({ onChange }) => {
   useEditorSubscriber();
 
-  const setSelectedNodeId = useEditorStore((state) => state.setSelectedNodeId);
   const dragItem = useDraggedItemStore((state) => state.componentId);
-  const updateItem = useDraggedItemStore((state) => state.updateItem);
   const setHoveringSection = useZoneStore((state) => state.setHoveringSection);
   const userIsDragging = !!dragItem;
   const breakpoints = useTreeStore((state) => state.breakpoints);
@@ -35,45 +28,14 @@ export const RootRenderer: React.FC<Props> = ({ onChange }) => {
   const { resolveDesignValue } = useBreakpoints(breakpoints);
   const tree = useTreeStore((state) => state.tree);
 
-  const { onAddComponent, onMoveComponent } = useCanvasInteractions();
   useEffect(() => {
     if (onChange) onChange(tree);
   }, [tree, onChange]);
 
-  const { onDragStartOrUpdate } = usePlaceholderStyle();
+  console.log(1);
 
   return (
-    <DragDropContext
-      onDragUpdate={(update) => {
-        updateItem(update);
-        onDragStartOrUpdate(update);
-      }}
-      onBeforeDragStart={(start) => {
-        onDragStartOrUpdate(start);
-        setSelectedNodeId('');
-        sendMessage(OUTGOING_EVENTS.ComponentSelected, {
-          nodeId: '',
-        });
-      }}
-      onDragEnd={(droppedItem) => {
-        updateItem(undefined);
-        dragState.reset();
-
-        sendMessage(OUTGOING_EVENTS.MouseUp);
-
-        // User cancel drag
-        if (!droppedItem.destination) {
-          sendMessage(OUTGOING_EVENTS.ComponentDragCanceled);
-          return;
-        }
-
-        // New component added to canvas
-        if (droppedItem.source.droppableId.startsWith('component-list')) {
-          onAddComponent(droppedItem);
-        } else {
-          onMoveComponent(droppedItem);
-        }
-      }}>
+    <DNDProvider>
       {dragItem && <DraggableContainer id={dragItem} />}
 
       <div className={styles.container}>
@@ -105,6 +67,6 @@ export const RootRenderer: React.FC<Props> = ({ onChange }) => {
           />
         )}
       </div>
-    </DragDropContext>
+    </DNDProvider>
   );
 };
