+++ packages/visual-editor/src/components/VisualEditorRoot.tsx
@@ -55,5 +55,7 @@ export const VisualEditorRoot = () => {
 
   if (!initialized) return null;
 
+  console.log('initialized');
+
   return <RootRenderer />;
 };
